+++ apps/www/src/app/(main)/(docs)/components/avatar/examples/nextjs-usage.tsx
@@ -4,11 +4,11 @@ import NextImage, { ImageProps } from "next/image";
 import Image from "~/avatar.png";
 
 // requires with UIConfigProvider
-declare module "@jamsr-ui/react" {
-  export interface AvatarProps extends ImageProps {
-    src: ImageProps["src"];
-  }
-}
+// declare module "@jamsr-ui/react" {
+//   export interface AvatarProps extends ImageProps {
+//     src: ImageProps["src"];
+//   }
+// }
 
 export const AvatarNextJsUsage = () => {
   return (
@@ -19,6 +19,7 @@ export const AvatarNextJsUsage = () => {
           isBordered: true,
         }}
       >
+        {/* @ts-ignore */}
         <Avatar alt="" src={Image} width={100} height={100} />
       </UIConfigProvider>
     </div>