+++ apps/www/src/app/(main)/(docs)/components/next/avatar/examples/global-config.tsx
@@ -1,0 +1,24 @@
+import { Avatar } from "@jamsr-ui/next";
+import { UIConfigProvider } from "@jamsr-ui/styles";
+import { useId } from "react";
+
+export const AvatarGlobalConfig = () => {
+  const id = useId();
+  return (
+    <UIConfigProvider
+      next={{
+        avatar: {
+          width: 100,
+          height: 100,
+          size: "lg",
+        },
+      }}
+    >
+      <Avatar
+        alt="image"
+        className="flex"
+        src={`https://i.pravatar.cc/300?u=${id}`}
+      />
+    </UIConfigProvider>
+  );
+};