+++ apps/www/src/app/(main)/(docs)/components/next/avatar/page.tsx
@@ -3,6 +3,7 @@ import { VariantWrapper } from "@/components/docs/variant-wrapper";
 import { type VariantTypes } from "@/types/variants";
 import { readVariantCode } from "@/utils/read-code";
 import { type Metadata } from "next";
+import { AvatarGlobalConfig } from "./examples/global-config";
 import { AvatarUsage } from "./examples/usage";
 
 const title = "Avatar";
@@ -14,7 +15,7 @@ export const metadata: Metadata = {
   description,
 };
 
-const code = <T extends VariantTypes["avatar"][number]>(variant: T) =>
+const code = <T extends VariantTypes["next-avatar"][number]>(variant: T) =>
   readVariantCode("avatar", variant, "next");
 
 const Avatar = () => {
@@ -22,6 +23,9 @@ const Avatar = () => {
     <VariantPage heading={title} description={description}>
       <VariantWrapper heading="Usage" code={code("usage")}>
         <AvatarUsage />
+      </VariantWrapper>
+      <VariantWrapper heading="Global Config" code={code("global-config")}>
+        <AvatarGlobalConfig />
       </VariantWrapper>
     </VariantPage>
   );