+++ apps/www/src/app/(main)/(docs)/components/next/link/examples/global-config.tsx
@@ -1,0 +1,18 @@
+import { Link } from "@jamsr-ui/next";
+import { UIConfigProvider } from "@jamsr-ui/styles";
+
+export const LinkGlobalConfig = () => {
+  return (
+    <UIConfigProvider
+      next={{
+        link: {
+          className: "text-red-500",
+          variant: "body4",
+          underline: "always",
+        },
+      }}
+    >
+      <Link href="/">Go to homepage </Link>
+    </UIConfigProvider>
+  );
+};