+++ apps/www/src/types/variants.d.ts
@@ -1,61 +1,385 @@
 export type VariantTypes = {
-  "accordion": ["custom-icon", "disabled", "end-content-outside", "end-content", "multiple-selection", "radius", "start-content-outside", "start-content", "subheading", "usage", "variants", "without-icon"];
-  "alert": ["custom-icon", "radius", "status", "usage", "variants", "with-action", "with-description", "without-icon"];
-  "autocomplete": ["choose-country", "controlled", "disabled-items", "disabled", "error-state", "helper-text", "multiple-custom-render", "multiple", "start-end-content", "usage"];
-  "avatar": ["bordered", "colors", "fallback-color", "fallbacks", "image-fallback", "nextjs-usage", "radius", "sizes", "usage"];
-  "breadcrumbs": [];
-  "button": ["colors", "disabled", "loading", "radius", "sizes", "usage", "variants-colors", "variants", "with-icons"];
-  "card": ["radius", "start-end-content", "usage", "variants", "with-form"];
-  "charts": [];
-  "checkbox": ["controlled", "description", "disabled", "invalid-state", "radius", "readonly", "usage"];
-  "chip": ["bordered", "colors", "customize", "radius", "sizes", "start-end-content", "usage", "variants-colors", "variants", "with-delete"];
-  "circular-progress": ["colors", "customization", "sizes", "usage", "with-label"];
-  "collapsible": ["controlled", "usage", "with-icon", "with-state"];
-  "confirmation": ["colors", "usage"];
-  "copy-to-clipboard": ["events", "global-config", "sound-effect", "usage", "with-children", "without-button"];
-  "data-table": ["bordered", "columns", "customization", "empty-state", "server", "sticky-header", "usage", "without-pagination"];
+  accordion: [
+    "custom-icon",
+    "disabled",
+    "end-content-outside",
+    "end-content",
+    "multiple-selection",
+    "radius",
+    "start-content-outside",
+    "start-content",
+    "subheading",
+    "usage",
+    "variants",
+    "without-icon",
+  ];
+  alert: [
+    "custom-icon",
+    "radius",
+    "status",
+    "usage",
+    "variants",
+    "with-action",
+    "with-description",
+    "without-icon",
+  ];
+  autocomplete: [
+    "choose-country",
+    "controlled",
+    "disabled-items",
+    "disabled",
+    "error-state",
+    "helper-text",
+    "multiple-custom-render",
+    "multiple",
+    "start-end-content",
+    "usage",
+  ];
+  avatar: [
+    "bordered",
+    "colors",
+    "fallback-color",
+    "fallbacks",
+    "image-fallback",
+    "nextjs-usage",
+    "radius",
+    "sizes",
+    "usage",
+  ];
+  breadcrumbs: [];
+  button: [
+    "colors",
+    "disabled",
+    "loading",
+    "radius",
+    "sizes",
+    "usage",
+    "variants-colors",
+    "variants",
+    "with-icons",
+  ];
+  card: ["radius", "start-end-content", "usage", "variants", "with-form"];
+  charts: [];
+  checkbox: [
+    "controlled",
+    "description",
+    "disabled",
+    "invalid-state",
+    "radius",
+    "readonly",
+    "usage",
+  ];
+  chip: [
+    "bordered",
+    "colors",
+    "customize",
+    "radius",
+    "sizes",
+    "start-end-content",
+    "usage",
+    "variants-colors",
+    "variants",
+    "with-delete",
+  ];
+  "circular-progress": [
+    "colors",
+    "customization",
+    "sizes",
+    "usage",
+    "with-label",
+  ];
+  collapsible: ["controlled", "usage", "with-icon", "with-state"];
+  confirmation: ["colors", "usage"];
+  "copy-to-clipboard": [
+    "events",
+    "global-config",
+    "sound-effect",
+    "usage",
+    "with-children",
+    "without-button",
+  ];
+  "data-table": [
+    "bordered",
+    "columns",
+    "customization",
+    "empty-state",
+    "server",
+    "sticky-header",
+    "usage",
+    "without-pagination",
+  ];
   "date-picker": [];
-  "dialog": ["backdrop", "bordered", "custom-close-button", "customize", "non-dismissible", "popover-items", "radius", "scroll-behavior", "usage", "without-animation", "without-close-button"];
-  "divider": ["colors", "customization", "orientation", "usage", "variants", "with-text"];
+  dialog: [
+    "backdrop",
+    "bordered",
+    "custom-close-button",
+    "customize",
+    "non-dismissible",
+    "popover-items",
+    "radius",
+    "scroll-behavior",
+    "usage",
+    "without-animation",
+    "without-close-button",
+  ];
+  divider: [
+    "colors",
+    "customization",
+    "orientation",
+    "usage",
+    "variants",
+    "with-text",
+  ];
   "drag-and-drop": ["disabled", "usage"];
-  "drawer": ["anchors", "backdrop", "bordered", "custom-close-button", "customize", "non-dismissible", "scroll-behavior", "size", "usage", "without-close-button"];
-  "editor": ["controlled", "invalid", "usage"];
+  drawer: [
+    "anchors",
+    "backdrop",
+    "bordered",
+    "custom-close-button",
+    "customize",
+    "non-dismissible",
+    "scroll-behavior",
+    "size",
+    "usage",
+    "without-close-button",
+  ];
+  editor: ["controlled", "invalid", "usage"];
   "empty-content": [];
-  "file-upload-multi": ["controlled", "customized", "default-value", "disabled", "radius", "usage"];
-  "file-upload-single": ["avatar", "controlled", "customized", "default-value", "disabled", "radius", "usage"];
-  "header": ["hide-on-scroll", "static", "usage", "with-border"];
-  "icon-button": ["colors", "disabled", "loading", "radius", "sizes", "usage", "variants-colors", "variants"];
-  "input": ["bordered-all-props", "clearable", "controlled", "disabled", "error-state", "filled", "helper-text", "label-helper", "numbers-only", "outlined-all-props", "placeholder", "radius", "required-optional", "secured-text", "sizes", "standard-all-props", "start-end-content", "underlined-all-props", "usage", "variants", "with-icon", "with-placeholder", "without-label"];
-  "kbd": ["keys", "radius", "usage"];
+  "file-upload-multi": [
+    "controlled",
+    "customized",
+    "default-value",
+    "disabled",
+    "radius",
+    "usage",
+  ];
+  "file-upload-single": [
+    "avatar",
+    "controlled",
+    "customized",
+    "default-value",
+    "disabled",
+    "radius",
+    "usage",
+  ];
+  header: ["hide-on-scroll", "static", "usage", "with-border"];
+  "icon-button": [
+    "colors",
+    "disabled",
+    "loading",
+    "radius",
+    "sizes",
+    "usage",
+    "variants-colors",
+    "variants",
+  ];
+  input: [
+    "bordered-all-props",
+    "clearable",
+    "controlled",
+    "disabled",
+    "error-state",
+    "filled",
+    "helper-text",
+    "label-helper",
+    "numbers-only",
+    "outlined-all-props",
+    "placeholder",
+    "radius",
+    "required-optional",
+    "secured-text",
+    "sizes",
+    "standard-all-props",
+    "start-end-content",
+    "underlined-all-props",
+    "usage",
+    "variants",
+    "with-icon",
+    "with-placeholder",
+    "without-label",
+  ];
+  kbd: ["keys", "radius", "usage"];
   "linear-progress": ["colors", "controlled", "usage"];
-  "link": ["nextjs-usage", "underline", "usage"];
-  "menu": ["backdrop", "colors", "controlled", "customize", "nested", "offset", "placement", "prevent-close", "radius", "start-end-content", "trigger-on-hover", "usage", "with-arrow", "without-lock-scroll"];
-  "next": [];
+  link: ["nextjs-usage", "underline", "usage"];
+  menu: [
+    "backdrop",
+    "colors",
+    "controlled",
+    "customize",
+    "nested",
+    "offset",
+    "placement",
+    "prevent-close",
+    "radius",
+    "start-end-content",
+    "trigger-on-hover",
+    "usage",
+    "with-arrow",
+    "without-lock-scroll",
+  ];
+  next: [];
   "number-input": [];
-  "otp-input": ["6-digits", "controlled", "numbers-only", "on-complete", "radius", "usage"];
-  "pagination": [];
-  "popover": ["controlled", "placement", "radius", "trigger-on", "usage", "with-arrow", "without-lock-scroll"];
-  "radio": ["checked", "colors", "disabled", "invalid", "radio-group-controlled", "radio-group-custom", "radio-group", "sizes", "usage"];
-  "rating": ["controlled", "default-value", "disabled", "helper-text", "invalid", "readonly", "size", "usage"];
-  "react-hook-form": ["autocomplete-multi", "autocomplete", "checkbox", "editor", "file-upload-multi-default-value", "file-upload-multi", "file-upload-single-default-value", "file-upload-single", "input", "otp-input", "radio-group", "rating", "select-multi", "select", "switch", "tags-input", "textarea"];
-  "repeater": ["index", "usage"];
-  "ripple": ["center", "usage"];
-  "select": ["controlled", "custom-render-complex", "custom-render-multi-search", "custom-render-multi", "custom-render-value", "disabled-items", "disabled", "invalid-state", "multiple-controlled", "multiple", "placeholder", "radius", "sizes", "start-end-items", "usage", "with-helper-text", "without-label"];
-  "sidebar": ["customized", "header-footer", "nested", "usage"];
-  "skeleton": ["usage"];
-  "slider": [];
-  "stepper": ["controlled", "events", "min-max-value", "usage"];
-  "switch": ["colors", "controlled", "description", "disabled", "invalid-state", "label-placement", "readonly", "sizes", "usage", "with-label"];
-  "tab": ["as", "colors", "customized", "disabled-item", "disabled", "outside", "radius", "sizes", "usage", "variants", "with-icons"];
-  "table": ["allow-hover", "bordered", "empty-state", "radius", "usage"];
-  "tags-input": ["controlled", "default-value", "invalid-state", "radius", "usage"];
-  "text": ["all", "usage"];
-  "textarea": ["controlled", "disabled", "error-state", "helper-text", "label-helper", "placeholder", "radius", "sizes", "usage", "variants", "with-end-content", "with-icon", "with-placeholder", "with-start-content", "without-label"];
-  "toast": ["radius", "usage"];
-  "tooltip": ["offset", "radius", "usage", "with-arrow"];
-  "area-chart": ["area-chart-gradient", "area-chart-multi-series-gradient", "area-chart-multi-series-stack-gradient", "area-chart-multi-series-stack", "area-chart-multi-series", "area-chart-solid"];
-  "bar-chart": ["bar-chart-custom", "bar-chart-multi-columns", "bar-chart-one-column", "bar-chart-stack", "bar-chart-vertical-multi-columns", "bar-chart-vertical-one-column", "bar-chart-vertical-stack"];
+  "otp-input": [
+    "6-digits",
+    "controlled",
+    "numbers-only",
+    "on-complete",
+    "radius",
+    "usage",
+  ];
+  pagination: [];
+  popover: [
+    "controlled",
+    "placement",
+    "radius",
+    "trigger-on",
+    "usage",
+    "with-arrow",
+    "without-lock-scroll",
+  ];
+  radio: [
+    "checked",
+    "colors",
+    "disabled",
+    "invalid",
+    "radio-group-controlled",
+    "radio-group-custom",
+    "radio-group",
+    "sizes",
+    "usage",
+  ];
+  rating: [
+    "controlled",
+    "default-value",
+    "disabled",
+    "helper-text",
+    "invalid",
+    "readonly",
+    "size",
+    "usage",
+  ];
+  "react-hook-form": [
+    "autocomplete-multi",
+    "autocomplete",
+    "checkbox",
+    "editor",
+    "file-upload-multi-default-value",
+    "file-upload-multi",
+    "file-upload-single-default-value",
+    "file-upload-single",
+    "input",
+    "otp-input",
+    "radio-group",
+    "rating",
+    "select-multi",
+    "select",
+    "switch",
+    "tags-input",
+    "textarea",
+  ];
+  repeater: ["index", "usage"];
+  ripple: ["center", "usage"];
+  select: [
+    "controlled",
+    "custom-render-complex",
+    "custom-render-multi-search",
+    "custom-render-multi",
+    "custom-render-value",
+    "disabled-items",
+    "disabled",
+    "invalid-state",
+    "multiple-controlled",
+    "multiple",
+    "placeholder",
+    "radius",
+    "sizes",
+    "start-end-items",
+    "usage",
+    "with-helper-text",
+    "without-label",
+  ];
+  sidebar: ["customized", "header-footer", "nested", "usage"];
+  skeleton: ["usage"];
+  slider: [];
+  stepper: ["controlled", "events", "min-max-value", "usage"];
+  switch: [
+    "colors",
+    "controlled",
+    "description",
+    "disabled",
+    "invalid-state",
+    "label-placement",
+    "readonly",
+    "sizes",
+    "usage",
+    "with-label",
+  ];
+  tab: [
+    "as",
+    "colors",
+    "customized",
+    "disabled-item",
+    "disabled",
+    "outside",
+    "radius",
+    "sizes",
+    "usage",
+    "variants",
+    "with-icons",
+  ];
+  table: ["allow-hover", "bordered", "empty-state", "radius", "usage"];
+  "tags-input": [
+    "controlled",
+    "default-value",
+    "invalid-state",
+    "radius",
+    "usage",
+  ];
+  text: ["all", "usage"];
+  textarea: [
+    "controlled",
+    "disabled",
+    "error-state",
+    "helper-text",
+    "label-helper",
+    "placeholder",
+    "radius",
+    "sizes",
+    "usage",
+    "variants",
+    "with-end-content",
+    "with-icon",
+    "with-placeholder",
+    "with-start-content",
+    "without-label",
+  ];
+  toast: ["radius", "usage"];
+  tooltip: ["offset", "radius", "usage", "with-arrow"];
+  "area-chart": [
+    "area-chart-gradient",
+    "area-chart-multi-series-gradient",
+    "area-chart-multi-series-stack-gradient",
+    "area-chart-multi-series-stack",
+    "area-chart-multi-series",
+    "area-chart-solid",
+  ];
+  "bar-chart": [
+    "bar-chart-custom",
+    "bar-chart-multi-columns",
+    "bar-chart-one-column",
+    "bar-chart-stack",
+    "bar-chart-vertical-multi-columns",
+    "bar-chart-vertical-one-column",
+    "bar-chart-vertical-stack",
+  ];
   "line-chart": ["line-chart-multi-series", "line-chart-simple"];
   "pie-chart": ["pie-chart-1", "pie-chart-2"];
-  "radial-bar-chart": ["radial-gauge", "radial-multi-bar", "radial-semi-circle-gauge", "radial-simple"];
+  "radial-bar-chart": [
+    "radial-gauge",
+    "radial-multi-bar",
+    "radial-semi-circle-gauge",
+    "radial-simple",
+  ];
+  "next-avatar": ["global-config", "usage"];
+  "next-link": ["global-config", "underline", "usage"];
 };