+++ packages/frameworks/next/src/types.d.ts
@@ -1,0 +1,11 @@
+import { type AvatarProps } from "./avatar";
+import { LinkProps } from "./link";
+
+declare module "@jamsr-ui/styles" {
+  export interface UIStylesType {
+    next?: {
+      avatar?: Partial<AvatarProps>;
+      link?: Partial<LinkProps>;
+    };
+  }
+}