+++ packages/components/rhf/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @jamsr-ui/rhf
 
+## 0.1.4
+
+### Patch Changes
+
+- Updated dependencies
+  - @jamsr-ui/editor@0.1.3
+
 ## 0.1.3
 
 ### Patch Changes